requirements.txt
aiohttp
beautifulsoup4
fake-useragent
fastapi
html5lib
loguru
lxml
pandas
playwright
psycopg2-binary
pydantic
python-dotenv
requests
selenium
sqlalchemy
streamlit
tqdm
uvicorn